+++ library/src/main/java/com/pengrad/telegrambot/request/ReopenForumTopic.java
@@ -4,11 +4,11 @@ import com.pengrad.telegrambot.response.BaseResponse;
 
 public class ReopenForumTopic extends BaseRequest<ReopenForumTopic, BaseResponse> {
 
-    public ReopenForumTopic(Long chatId, Integer messageThreadId) {
+    public ReopenForumTopic(Long chatId, Long messageThreadId) {
         this(chatId.toString(), messageThreadId);
     }
 
-    public ReopenForumTopic(String chatId, Integer messageThreadId) {
+    public ReopenForumTopic(String chatId, Long messageThreadId) {
         super(BaseResponse.class);
         add("chat_id", chatId);
         add("message_thread_id", messageThreadId);
