+++ library/src/main/java/com/pengrad/telegrambot/request/UnpinAllForumTopicMessages.java
@@ -4,11 +4,11 @@ import com.pengrad.telegrambot.response.BaseResponse;
 
 public class UnpinAllForumTopicMessages extends BaseRequest<UnpinAllForumTopicMessages, BaseResponse> {
 
-    public UnpinAllForumTopicMessages(Long chatId, Integer messageThreadId) {
+    public UnpinAllForumTopicMessages(Long chatId, Long messageThreadId) {
         this(chatId.toString(), messageThreadId);
     }
 
-    public UnpinAllForumTopicMessages(String chatId, Integer messageThreadId) {
+    public UnpinAllForumTopicMessages(String chatId, Long messageThreadId) {
         super(BaseResponse.class);
         add("chat_id", chatId);
         add("message_thread_id", messageThreadId);
